+++ app/va/va_profile/va_profile_client.py
@@ -114,61 +114,6 @@ class VAProfileClient:
         response_json: dict = response.json()
         return response_json.get('profile', {})
 
-    def get_telephone(self, va_profile_id: RecipientIdentifier) -> str:
-        """
-        Retrieve the telephone number from the profile information for a given VA profile ID.
-
-        Args:
-            va_profile_id (RecipientIdentifier): The VA profile ID to retrieve the telephone number for.
-
-        Returns:
-            str: The telephone number retrieved from the VA Profile service.
-        """
-        contact_info: ContactInformation = self.get_profile(va_profile_id).get('contactInformation', {})
-        self.logger.debug('V3 Profile - Retrieved ContactInformation: %s', contact_info)
-
-        telephones: list[Telephone] = contact_info.get(self.PHONE_BIO_TYPE, [])
-        sorted_telephones = sorted(
-            [phone for phone in telephones if phone['phoneType'] == PhoneNumberType.MOBILE.value],
-            key=lambda phone: iso8601.parse_date(phone['createDate']),
-            reverse=True,
-        )
-        if sorted_telephones:
-            if (
-                sorted_telephones[0].get('countryCode')
-                and sorted_telephones[0].get('areaCode')
-                and sorted_telephones[0].get('phoneNumber')
-            ):
-                self.statsd_client.incr('clients.va-profile.get-telephone.success')
-                # https://en.wikipedia.org/wiki/E.164 format
-                return f"+{sorted_telephones[0]['countryCode']}{sorted_telephones[0]['areaCode']}{sorted_telephones[0]['phoneNumber']}"
-
-        self.statsd_client.incr('clients.va-profile.get-telephone.failure')
-        self._raise_no_contact_info_exception(self.PHONE_BIO_TYPE, va_profile_id, contact_info.get(self.TX_AUDIT_ID))
-
-    def get_email(self, va_profile_id: RecipientIdentifier) -> str:
-        """
-        Retrieve the email address from the profile information for a given VA profile ID.
-
-        Args:
-            va_profile_id (RecipientIdentifier): The VA profile ID to retrieve the email address for.
-
-        Returns:
-            str: The email address retrieved from the VA Profile service.
-        """
-        contact_info: ContactInformation = self.get_profile(va_profile_id).get('contactInformation', {})
-        sorted_emails = sorted(
-            contact_info.get(self.EMAIL_BIO_TYPE, []),
-            key=lambda email: iso8601.parse_date(email['createDate']),
-            reverse=True,
-        )
-        if sorted_emails:
-            self.statsd_client.incr('clients.va-profile.get-email.success')
-            return sorted_emails[0].get('emailAddressText')
-
-        self.statsd_client.incr('clients.va-profile.get-email.failure')
-        self._raise_no_contact_info_exception(self.EMAIL_BIO_TYPE, va_profile_id, contact_info.get(self.TX_AUDIT_ID))
-
     def has_valid_mobile_telephone_classification(self, telephone: Telephone, contact_info: ContactInformation):
         """
         Args:
@@ -223,7 +168,7 @@ class VAProfileClient:
                 self.statsd_client.incr('clients.va-profile.get-telephone.success')
                 return f"+{sorted_telephones[0]['countryCode']}{sorted_telephones[0]['areaCode']}{sorted_telephones[0]['phoneNumber']}"
 
-    def get_telephone_with_permission(
+    def get_telephone(
         self,
         va_profile_id: RecipientIdentifier,
         notification: Notification,
@@ -263,7 +208,7 @@ class VAProfileClient:
 
         return VAProfileResult(telephone, communication_allowed, permission_message)
 
-    def get_email_with_permission(
+    def get_email(
         self,
         va_profile_id: RecipientIdentifier,
         notification: Notification,
@@ -280,6 +225,9 @@ class VAProfileClient:
             Property recipient is the telephone number retrieved from the VA Profile service.
             Property communication_allowed is true when VA Profile service indicates that the recipient has allowed communication.
             Property permission_message may contain an error message if the permission check encountered an exception.
+
+        Raises:
+            NoContactInfoException: if an email address is not found for the given VA profile ID.
         """
         profile = self.get_profile(va_profile_id)
         communication_allowed = notification.default_send
@@ -306,61 +254,6 @@ class VAProfileClient:
 
         self.statsd_client.incr('clients.va-profile.get-email.failure')
         self._raise_no_contact_info_exception(self.EMAIL_BIO_TYPE, va_profile_id, contact_info.get(self.TX_AUDIT_ID))
-
-    def get_is_communication_allowed(
-        self,
-        recipient_id: RecipientIdentifier,
-        communication_item_id: str,
-        notification_id: str,
-        notification_type: str,
-        default_send: bool,
-    ) -> bool:
-        """
-        Determine if communication is allowed for a given recipient, communication item, and notification type.
-
-        Args:
-            recipient_id (RecipientIdentifier): The recipient's VA profile ID.
-            communication_item_id (str): The ID of the communication item.
-            notification_id (str): The ID of the notification.
-            notification_type (str): The type of the notification.
-
-        Returns:
-            bool: True if communication is allowed, False otherwise.
-
-        Raises:
-            CommunicationItemNotFoundException: If no communication permissions are found for the given parameters.
-        """
-
-        communication_permissions: CommunicationPermissions = self.get_profile(recipient_id).get(
-            'communicationPermissions', {}
-        )
-        communication_channel = CommunicationChannel(VA_NOTIFY_TO_VA_PROFILE_NOTIFICATION_TYPES[notification_type])
-        for perm in communication_permissions:
-            if (
-                perm['communicationChannelId'] == communication_channel.id
-                and perm['communicationItemId'] == communication_item_id
-            ):
-                self.statsd_client.incr('clients.va-profile.get-communication-item-permission.success')
-                # if default send is true and allowed is false, return false
-                # if default send is true and allowed is true, return true
-                # if default send is false, default to what it finds
-                permission: bool | None = perm['allowed']
-                if permission is not None:
-                    return perm['allowed']
-                else:
-                    return default_send
-
-        self.logger.debug(
-            'V3 Profile -- Recipient %s did not have permission for communication item %s and channel %s for notification %s',
-            recipient_id,
-            communication_item_id,
-            notification_type,
-            notification_id,
-        )
-
-        # TODO 893 - use default communication item settings when that has been implemented
-        self.statsd_client.incr('clients.va-profile.get-communication-item-permission.no-permissions')
-        raise CommunicationItemNotFoundException
 
     def get_is_communication_allowed_from_profile(
         self,
@@ -470,13 +363,6 @@ class VAProfileClient:
         self.statsd_client.incr(f'clients.va-profile.get-{bio_type}.no-{bio_type}')
         raise NoContactInfoException(
             f'No {bio_type} in response for VA Profile ID {va_profile_id} ' f'with AuditId {tx_audit_id}'
-        )
-
-    def _raise_invalid_phone_number_exception(self, contact_info: ContactInformation):
-        self.statsd_client.incr(f'clients.va-profile.get-{self.PHONE_BIO_TYPE}.no-{self.PHONE_BIO_TYPE}')
-        raise InvalidPhoneNumberException(
-            f'No valid {self.PHONE_BIO_TYPE} in response for VA Profile ID {contact_info.get("vaProfileId")} '
-            f'with AuditId {contact_info.get("txAuditId")}'
         )
 
     def send_va_profile_email_status(self, notification_data: dict) -> None:
